Process Quick Select, Scan Optimizations and Bug Fixes

+++ Ana/Source/Engine/OperatingSystems/NormalizedModule.cs
@@ -3,17 +3,17 @@
     using System;
 
     /// <summary>
-    /// Defines an OS independent module region and attributes
+    /// Defines an OS independent module region and attributes.
     /// </summary>
     internal class NormalizedModule : NormalizedRegion
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="NormalizedModule" /> class
+        /// Initializes a new instance of the <see cref="NormalizedModule" /> class.
         /// </summary>
-        /// <param name="name">The name of the module</param>
-        /// <param name="baseAddress">The base address of the module</param>
-        /// <param name="size">The total size of the module</param>
-        public NormalizedModule(String name, IntPtr baseAddress, Int32 size) : base(baseAddress, size)
+        /// <param name="name">The name of the module.</param>
+        /// <param name="baseAddress">The base address of the module.</param>
+        /// <param name="size">The total size of the module.</param>
+        public NormalizedModule(String name, IntPtr baseAddress, UInt64 size) : base(baseAddress, size)
         {
             this.Name = name;
         }
